+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 8.0.0
 
+### Enhancements
+- **General Changes**
+  - Added a page to track consent revisions
+
 ## 7.0.0 (February 11, 2018)
 
 ### Enhancements